requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.5
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.8.1
astroid==3.0.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
auth0-python==4.7.2
Automat==20.2.0
axios==0.4.0
Babel==2.8.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.4
bs4==0.0.1
build==1.2.2.post1
cachetools==5.5.0
certifi==2023.11.17
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
cloud-init==22.2
cohere==5.11.4
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
crewai==0.80.0
crewai-tools==0.14.0
cryptography==43.0.3
dataclasses-json==0.6.7
dbus-python==1.2.18
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.7
distro==1.7.0
distro-info===1.1build1
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
durationpy==0.9
embedchain==0.1.125
exceptiongroup==1.2.0
executing==2.1.0
fastapi==0.115.5
fastavro==1.9.7
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-aiplatform==1.72.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5validator==0.4.2
httpcore==1.0.7
httpie==2.6.0
httplib2==0.20.2
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.5.0
importlib_resources==6.4.5
incremental==21.3.0
iniconfig==2.0.0
instructor==1.6.4
ipython==8.29.0
isort==5.13.2
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.6.1
json_repair==0.30.2
jsonpatch==1.33
jsonpickle==4.0.0
jsonpointer==2.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
kubernetes==31.0.0
lancedb==0.16.0
langchain==0.3.7
langchain-cohere==0.3.1
langchain-community==0.3.7
langchain-core==0.3.19
langchain-experimental==0.3.3
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langsmith==0.1.143
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
litellm==1.52.10
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.23.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mem0ai==0.1.29
mmh3==5.0.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.0
openai==1.54.5
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
optional-django==0.3.0
orjson==3.10.11
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pexpect==4.8.0
platformdirs==4.1.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.2
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydocstyle==6.3.0
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.10.0
pylance==0.19.2
pylint==3.0.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.389
pyrsistent==0.18.1
pysbd==0.3.4
pyserial==3.5
PySocks==1.7.1
pytest==8.3.3
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.9.0.post0
python-debian===0.1.43ubuntu1
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.1
qdrant-client==1.12.1
react==4.3.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
schema==0.7.7
SecretStorage==3.3.1
selenium==4.26.1
service-identity==18.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sos==4.3
soupsieve==2.5
SQLAlchemy==2.0.35
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
systemd-python==234
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.20.3
tomli==2.1.0
tomli_w==1.1.0
tomlkit==0.12.3
tqdm==4.67.0
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
Twisted==22.1.0
typer==0.13.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-advantage-tools==27.9
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
uv==0.5.2
uvicorn==0.32.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.2
zipp==3.21.0
zope.interface==5.4.0
